MOD / 3-3-1. 남은 시간 확인 / SSE 연결 완료, 메시지 테스트 중

+++ springboot/src/main/java/com/petmeeting/springboot/domain/Chat.java
@@ -30,5 +30,5 @@ public class Chat {
     private String content;
 
     @Column(name = "created_time", nullable = false)
-    private Integer createdTime;
+    private Long createdTime;
 }
